PostgreSQL support added.

+++ read_users.php
@@ -3,32 +3,39 @@
     require_once 'config.inc.php';
     
     readfile('header.tmpl.html');
+
+    require 'connections/postgresql.php';
 ?>
 
 <div class="container p-3 my-3 border">
 <h1 class="text-center">List of users</h1>
 <ul class="list-group">
 <?php
-    $db = new mysqli(MYSQL_HOST, MYSQL_USER, MYSQL_PASSWORD, MYSQL_DATABASE);
-    $sql = 'SELECT * FROM users WHERE isAdmin=0';
-    $result = $db->query($sql);
+    $db = new PostgreSQLDBContext();
+    $db->query("SELECT * FROM users WHERE isAdmin=b'0'");
     
-    foreach ($result as $row) {
-        printf(
-            '<li class="list-group-item d-flex justify-content-between"><span style="color: %s">%s (%s)</span>
-                <div>
-                    <a href="update_user.php?id=%s" class="btn btn-sm btn-primary active" role="button">Update</a>
-                    <a href="delete_user.php?id=%s" class="btn btn-sm btn-danger active" role="button">Delete</a>
-                </div>
-            </h1>',
-            htmlspecialchars($row['color'], ENT_QUOTES),
-            htmlspecialchars($row['name'], ENT_QUOTES),
-            htmlspecialchars($row['gender'], ENT_QUOTES),
-            htmlspecialchars($row['id'], ENT_QUOTES),
-            htmlspecialchars($row['id'], ENT_QUOTES));
-    };
-
-    $db->close();
+    $lines = $db->fetch_all_results();
+    if ($lines != null && $lines != false) {
+        foreach ($lines as $line) {
+            printf(
+                '<li class="list-group-item d-flex justify-content-between"><span style="color: %s">%s (%s)</span>
+                    <div>
+                        <a href="update_user.php?id=%s" class="btn btn-sm btn-primary active" role="button">Update</a>
+                        <a href="delete_user.php?id=%s" class="btn btn-sm btn-danger active" role="button">Delete</a>
+                    </div>
+                </h1>',
+                htmlspecialchars($line['color'], ENT_QUOTES),
+                htmlspecialchars($line['name'], ENT_QUOTES),
+                htmlspecialchars($line['gender'], ENT_QUOTES),
+                htmlspecialchars($line['id'], ENT_QUOTES),
+                htmlspecialchars($line['id'], ENT_QUOTES));
+        };
+    } else {
+        echo '<div class="text-info">
+        <p class="text-center">Looks like there are no Users on the Database
+        <br>Try to create a new user, to see what happens.</p></div>';
+    }
+    
 ?>
 </ul>
 <br>
